requirements.txt
numpy==1.25.1
pandas==2.0.3
python-dotenv==1.0.1
